+++ server/agents/nightlyCostAlert/tools.js
@@ -1,0 +1,6 @@
+'use strict';
+
+const TOOL_SLUG = 'nightly-cost-alert';
+const nightlyCostAlertTools = [];
+
+module.exports = { nightlyCostAlertTools, TOOL_SLUG };